requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.2
Flask-Migrate==4.0.3
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
email-validator==1.3.1
Werkzeug==2.3.2
WTForms==3.0.1
gunicorn==20.1.0
pymysql==1.0.2
python-dotenv==1.0.0
cryptography==41.0.2
itsdangerous==2.1.2
